# api/openapi.bundled.yaml
openapi: 3.2.0
...
info:
  title: Arches Platform API
  description: The Arches Platform API for managing projects and code generation
  version: v0.0.0
  summary: API for Arches platform
  termsOfService: https://archesai.com/terms
  license:
    name: AGPL-3.0
    url: https://opensource.org/licenses/AGPL-3.0
  contact:
    name: Arches AI Support
    email: support@archesai.com
    url: https://archesai.com/support
...
servers:
  - url: https://api.archesai.dev
    description: Remote development server
  - url: https://api.archesai.com
paths:
...
  /auth/sessions/{id}:
    get:
      x-internal: auth
      operationId: GetSession
      summary: Find a session
      tags:
        - Session
      description: Find a session
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      x-internal: auth
      x-codegen-custom-handler: true
      operationId: UpdateSession
      summary: Update Session
      tags:
        - Auth
      description: This endpoint will update the active organization for the current session
      requestBody:
        description: Session update data
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationID:
                  description: The organization ID to set as active for this session
                  type: string
                  format: uuid
                  example: 550e8400-e29b-41d4-a716-446655440000
                  minLength: 36
                  maxLength: 36
              required:
                - organizationID
        required: true
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          $ref: '#/components/responses/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      x-internal: auth
      x-codegen-custom-handler: true
      operationId: DeleteSession
      summary: Delete session (Logout)
      tags:
        - Auth
      description: Delete a specific session (logout from that session)
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{id}:
    get:
      x-internal: auth
      operationId: GetUser
      summary: Get a user
      tags:
        - User
      description: Get a user
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      x-internal: auth
      operationId: UpdateUser
      summary: Update an user
      tags:
        - User
      description: Update an user
      requestBody:
        description: User update data
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  description: The user's e-mail
                  type: string
                  format: email
                  minLength: 1
                  example: user@example.com
                  maxLength: 255
                image:
                  description: The user's avatar image URL
                  type: string
                  format: uri
                  maxLength: 2048
                  example: https://example.com/avatar.jpg
                  pattern: ^https?://[\w\-\.]+[\w\-\./:?#@!$&'()*+,;=]*$
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      x-internal: auth
      operationId: DeleteUser
      summary: Delete a user
      tags:
        - User
      description: Delete a user
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      security:
        - bearerAuth: []
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Session:
      x-codegen-schema-type: entity
      x-codegen:
        repository:
          excludeFromUpdate:
            - IPAddress
            - UserID
            - UserAgent
            - Token
          relations:
            - field: userID
              references: user
              onDelete: CASCADE
              onUpdate: NO_ACTION
          indices:
            - userID
            - token
      x-internal: auth
      description: Schema for Session entity
      title: Session
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          properties:
            organizationID:
              description: The organization ID for this session (nullable for users without org)
              type:
                - string
                - 'null'
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
              minLength: 36
              maxLength: 36
            authMethod:
              description: The authentication method used (magic_link, oauth_google, oauth_github, etc.)
              type:
                - string
                - 'null'
              minLength: 1
              maxLength: 100
              example: magic_link
              pattern: ^[a-z_]+$
            authProvider:
              description: The authentication provider (google, github, microsoft, local)
              type:
                - string
                - 'null'
              enum:
                - local
                - google
                - github
                - microsoft
                - apple
              example: local
            expiresAt:
              description: The expiration date of the session
              type: string
              minLength: 1
              format: date-time
              example: '2023-10-05T14:48:00Z'
              maxLength: 255
            ipAddress:
              description: The IP address of the session
              type:
                - string
                - 'null'
              format: ipv4
              example: 192.168.1.1
              minLength: 7
              maxLength: 45
            token:
              description: The session token
              type: string
              minLength: 1
              example: eyJhbGciOiJIUzI1NiIs...
              maxLength: 512
              pattern: ^[A-Za-z0-9_\-\.]+$
            userAgent:
              description: The user agent of the session
              type:
                - string
                - 'null'
              example: Mozilla/5.0 (Windows NT 10.0; Win64; x64)
              minLength: 1
              maxLength: 255
              pattern: ^[\w\s\-.,!?()@#+/\(\):;]+$
            userID:
              description: The user who owns this session
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
              minLength: 36
              maxLength: 36
          required:
            - expiresAt
            - token
            - userID
            - organizationID
            - authMethod
            - authProvider
            - ipAddress
            - userAgent
      unevaluatedProperties: false
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        createdAt: '2023-10-05T14:48:00Z'
        updatedAt: '2023-10-05T14:48:00Z'
        organizationID: 550e8400-e29b-41d4-a716-446655440000
        authMethod: magic_link
        authProvider: local
        expiresAt: '2023-10-05T14:48:00Z'
        ipAddress:
    User:
      description: Schema for User entity
      title: User
      x-codegen-schema-type: entity
      x-codegen:
        repository:
          indices:
            - email
          additionalMethods:
            - name: GetUserByEmail
              params:
                - name: email
                  type: string
              returns: single
            - name: GetUserBySessionID
              params:
                - name: sessionID
                  type: string
                  format: uuid
                  maxLength: 36
              returns: single
      x-internal: auth
      allOf:
        - $ref: '#/components/schemas/Base'
        - type: object
          properties:
            email:
              description: The user's email address
              type: string
              format: email
              minLength: 5
              maxLength: 255
              example: user@example.com
            emailVerified:
              description: Whether the user's email has been verified
              type: boolean
              default: false
              example: true
            image:
              description: The user's avatar image URL
              type:
                - string
                - 'null'
              minLength: 5
              maxLength: 2048
              format: uri
              example: https://example.com/avatar.jpg
            name:
              description: The user's display name
              type: string
              minLength: 1
              maxLength: 255
              example: John Doe
              pattern: ^[\w\s\-.,!?()@#+/']+$
          required:
            - email
            - emailVerified
            - name
            - image
      unevaluatedProperties: false
...
    Base:
      title: Base
      description: Base schema for all entities with common fields
      x-internal: server
      type: object
      properties:
        id:
          description: Unique identifier for the resource
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          minLength: 36
          maxLength: 36
        createdAt:
          description: The date and time when the resource was created
          type: string
          format: date-time
          example: '2024-01-15T09:30:00Z'
          minLength: 1
          maxLength: 255
        updatedAt:
          description: The date and time when the resource was last updated
          type: string
          format: date-time
          example: '2024-01-15T10:45:00Z'
          minLength: 1
          maxLength: 255
      required:
        - id
        - createdAt
        - updatedAt
...
    UUID:
      description: UUID identifier
      x-internal: server
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
      minLength: 36
      maxLength: 36
...
    Problem:
      title: Problem
      description: RFC 7807 (Problem Details) compliant error response
      x-internal: server
      type: object
      additionalProperties: false
      required:
        - status
        - title
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code
          minimum: 100
          maximum: 599
          example: 400
        type:
          type: string
          format: uri
          description: URI identifying the problem type
          default: about:blank
          example: https://api.example.com/errors/validation-failed
          minLength: 1
          maxLength: 2048
        title:
          type: string
          description: Short, human-readable summary
          example: Validation Failed
          minLength: 1
          maxLength: 255
          pattern: ^[\w\s\-.,!?()]+$
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: The request body contains invalid fields
          minLength: 1
          maxLength: 255
          pattern: ^[\w\s\-.,!?()]+$
        instance:
          type: string
          format: uri
          description: URI identifying the specific occurrence
          example: https://api.example.com/auth/users/123
          maxLength: 2048
...
  parameters:
...
    ResourceID:
      schema:
        $ref: '#/components/schemas/UUID'
      in: path
      name: id
      required: true
      description: The unique identifier of the resource.
...
  responses:
    BadRequest:
      description: 400 Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    NoContent:
      description: 204 No Content
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    NotFound:
      description: 404 Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
    Unauthorized:
      description: 401 Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
    UnprocessableEntity:
      description: 422 Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    SessionResponse:
      description: Session retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Session'
            required:
              - data
            additionalProperties: false
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
    UserResponse:
      description: User retrieved successfully
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/User'
            required:
              - data
            additionalProperties: false
          example:
            data:
              id: 550e8400-e29b-41d4-a716-446655440000
              name: John Doe
              email: user@example.com
              emailVerified: true
              image: https://example.com/avatar.jpg
              createdAt: '2024-01-15T09:30:00Z'
              updatedAt: '2024-01-15T09:30:00Z'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
...
  headers:
    RateLimitLimit:
      description: The maximum number of requests allowed per time window
      schema:
        type: integer
        example: 1000
        format: int32
        minimum: 1
        maximum: 1000000
      example: 1000
    RateLimitRemaining:
      description: The number of requests remaining in the current time window
      schema:
        type: integer
        example: 999
        format: int32
        minimum: 0
        maximum: 1000000
      example: 999
    RateLimitReset:
      description: The time at which the current rate limit window resets (Unix timestamp)
      schema:
        type: integer
        example: 1640995200
        format: int64
        minimum: 0
        maximum: 9223372036854776000
      example: 1640995200
    RetryAfter:
      description: Number of seconds to wait before making a new request
      schema:
        type: integer
        example: 60
        format: int32
        minimum: 1
        maximum: 3600
      example: 60
...
  securitySchemes:
    bearerAuth:
      description: API Token for authenticated requests
      scheme: bearer
      type: http
    sessionCookie:
      description: Session cookie for authenticated requests
      type: apiKey
      name: __Secure-better-auth.session_token
      in: cookie
security:
  - bearerAuth: []
  - sessionCookie: []