# pkg/auth/gen/controllers/update_session.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"
	"github.com/oapi-codegen/runtime"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/auth/gen/models"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// UpdateSession - PATCH /auth/sessions/{id}
// ============================================================================

// UpdateSessionHandler is the HTTP handler for UpdateSession.
type UpdateSessionHandler struct {
	updateSession application.UpdateSession
}

// NewUpdateSessionHandler creates a new HTTP handler.
func NewUpdateSessionHandler(updateSession application.UpdateSession) *UpdateSessionHandler {
	return &UpdateSessionHandler{updateSession: updateSession}
}

// RegisterUpdateSessionRoute registers the HTTP route for UpdateSession.
func RegisterUpdateSessionRoute(mux *http.ServeMux, handler *UpdateSessionHandler) {
	mux.HandleFunc("PATCH /auth/sessions/{id}", handler.ServeHTTP)
}

// Request types

// UpdateSessionRequestBody defines the request body for UpdateSession
type UpdateSessionRequestBody struct {
	OrganizationID uuid.UUID `json:"organizationID"`
}

// Response types

type UpdateSessionResponse interface {
	VisitUpdateSessionResponse(w http.ResponseWriter) error
}

type UpdateSession200Response struct {
	Data models.Session `json:"data"`
}

func (response UpdateSession200Response) VisitUpdateSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	return json.NewEncoder(w).Encode(response)
}

type UpdateSession400Response struct {
	server.ProblemDetails
}

func (response UpdateSession400Response) VisitUpdateSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateSession401Response struct {
	server.ProblemDetails
}

func (response UpdateSession401Response) VisitUpdateSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateSession404Response struct {
	server.ProblemDetails
}

func (response UpdateSession404Response) VisitUpdateSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateSession422Response struct {
	server.ProblemDetails
}

func (response UpdateSession422Response) VisitUpdateSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateSession429Response struct {
	server.ProblemDetails
}

func (response UpdateSession429Response) VisitUpdateSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateSession500Response struct {
	server.ProblemDetails
}

func (response UpdateSession500Response) VisitUpdateSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the PATCH /auth/sessions/{id} endpoint.
func (h *UpdateSessionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := UpdateSession401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitUpdateSessionResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.UpdateSessionInput{}
	input.SessionID = sessionID

	// Path parameter "id"
	var id uuid.UUID
	if err := runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true}); err != nil {
		errorResp := UpdateSession400Response{
			ProblemDetails: server.NewBadRequestResponse(fmt.Sprintf("Invalid format for parameter id: %s", err), r.URL.Path),
		}
		if err := errorResp.VisitUpdateSessionResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.ID = id

	// Request body
	body := &UpdateSessionRequestBody{}
	if err := json.NewDecoder(r.Body).Decode(body); err != nil {
		errorResp := UpdateSession400Response{
			ProblemDetails: server.NewBadRequestResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitUpdateSessionResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.OrganizationID = body.OrganizationID

	// Execute
	result, err := h.updateSession.Execute(ctx, input)
	if err != nil {
		errorResp := UpdateSession500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitUpdateSessionResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := UpdateSession200Response{}
	response.Data = result.Data

	if err := response.VisitUpdateSessionResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}