# pkg/auth/gen/controllers/update_user.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"
	"github.com/oapi-codegen/runtime"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/auth/gen/models"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// UpdateUser - PATCH /users/{id}
// ============================================================================

// UpdateUserHandler is the HTTP handler for UpdateUser.
type UpdateUserHandler struct {
	updateUser application.UpdateUser
}

// NewUpdateUserHandler creates a new HTTP handler.
func NewUpdateUserHandler(updateUser application.UpdateUser) *UpdateUserHandler {
	return &UpdateUserHandler{updateUser: updateUser}
}

// RegisterUpdateUserRoute registers the HTTP route for UpdateUser.
func RegisterUpdateUserRoute(mux *http.ServeMux, handler *UpdateUserHandler) {
	mux.HandleFunc("PATCH /users/{id}", handler.ServeHTTP)
}

// Request types

// UpdateUserRequestBody defines the request body for UpdateUser
type UpdateUserRequestBody struct {
	Email *string `json:"email,omitempty"`
	Image *string `json:"image,omitempty"`
}

// Response types

type UpdateUserResponse interface {
	VisitUpdateUserResponse(w http.ResponseWriter) error
}

type UpdateUser200Response struct {
	Data models.User `json:"data"`
}

func (response UpdateUser200Response) VisitUpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	return json.NewEncoder(w).Encode(response)
}

type UpdateUser400Response struct {
	server.ProblemDetails
}

func (response UpdateUser400Response) VisitUpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateUser401Response struct {
	server.ProblemDetails
}

func (response UpdateUser401Response) VisitUpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateUser404Response struct {
	server.ProblemDetails
}

func (response UpdateUser404Response) VisitUpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateUser422Response struct {
	server.ProblemDetails
}

func (response UpdateUser422Response) VisitUpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateUser429Response struct {
	server.ProblemDetails
}

func (response UpdateUser429Response) VisitUpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type UpdateUser500Response struct {
	server.ProblemDetails
}

func (response UpdateUser500Response) VisitUpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the PATCH /users/{id} endpoint.
func (h *UpdateUserHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := UpdateUser401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitUpdateUserResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.UpdateUserInput{}
	input.SessionID = sessionID

	// Path parameter "id"
	var id uuid.UUID
	if err := runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true}); err != nil {
		errorResp := UpdateUser400Response{
			ProblemDetails: server.NewBadRequestResponse(fmt.Sprintf("Invalid format for parameter id: %s", err), r.URL.Path),
		}
		if err := errorResp.VisitUpdateUserResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.ID = id

	// Request body
	body := &UpdateUserRequestBody{}
	if err := json.NewDecoder(r.Body).Decode(body); err != nil {
		errorResp := UpdateUser400Response{
			ProblemDetails: server.NewBadRequestResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitUpdateUserResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.Email = body.Email
	input.Image = body.Image

	// Execute
	result, err := h.updateUser.Execute(ctx, input)
	if err != nil {
		errorResp := UpdateUser500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitUpdateUserResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Map output to response
	response := UpdateUser200Response{}
	response.Data = result.Data

	if err := response.VisitUpdateUserResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}